driver: src: Split devices into into different source files
Finally split devices into their own source files, no more  10 billion line files.
