+++ driver/src/ref/hobovr_components.h
@@ -7,6 +7,23 @@
 #ifndef HOBOVR_COMPONENTS_H
 #define HOBOVR_COMPONENTS_H
 
+#include <openvr_driver.h>
+#include <cmath>
+#include <memory>
+
+#if defined(_WIN32)
+
+#include "receiver_win.h"
+
+#elif defined(__linux__)
+
+#include "receiver_linux.h"
+#define _stricmp strcasecmp
+
+#endif
+
+#include "hobovr_defines.h"
+
 namespace hobovr {
   enum ELensMathType {
     Mt_Invalid = 0,
@@ -177,9 +194,9 @@ namespace hobovr {
       }
     }
 
-    virtual DistortionCoordinates_t ComputeDistortion(vr::EVREye eEye, float fU,
+    virtual vr::DistortionCoordinates_t ComputeDistortion(vr::EVREye eEye, float fU,
                                                       float fV) {
-      DistortionCoordinates_t coordinates;
+      vr::DistortionCoordinates_t coordinates;
 
       (void)eEye; // because its not used and i cant remove this argument
