+++ driver/src/ref/util.h
@@ -16,7 +16,7 @@
 namespace SockReceiver {
   //can receive packets ending with \t\r\n using either winsock2 or unix sockets
   template <typename T>
-  int receive_till_zero( T sock, char* buf, int& numbytes, int max_packet_size )
+  inline int receive_till_zero( T sock, char* buf, int& numbytes, int max_packet_size )
   {
     // receives a message until an end token is reached
     // thanks to https://stackoverflow.com/a/13528453/10190971
@@ -53,7 +53,7 @@ namespace SockReceiver {
 
   // reshapes packet vector into shape inxs
   template <typename T>
-  std::vector<std::vector<T>> split_pk(std::vector<T> arr, std::vector<int> inxs) {
+  inline std::vector<std::vector<T>> split_pk(std::vector<T> arr, std::vector<int> inxs) {
     std::vector<std::vector<T>> out;
     int oof = 0;
     int ms = (int)arr.size();
@@ -73,16 +73,8 @@ namespace SockReceiver {
     return out;
   }
 
-  // returns the found text
-  std::string first_rgx_match(std::string ss, std::regex rgx) {
-    std::smatch mm;
-    if (std::regex_search(ss, mm, rgx))
-      return mm[0];
-    return "";
-  }
-
   // get regex vector
-  std::vector<std::string> get_rgx_vector(std::string ss, std::regex rgx) {
+  inline std::vector<std::string> get_rgx_vector(std::string ss, std::regex rgx) {
     std::sregex_token_iterator iter(ss.begin(),
         ss.end(),
         rgx,
@@ -95,32 +87,9 @@ namespace SockReceiver {
     return out;
   }
 
-  // split by regex
-  std::vector<std::string> split_by_rgx(std::string ss, std::regex rgx) {
-    std::sregex_token_iterator iter(ss.begin(),
-        ss.end(),
-        rgx,
-        -1);
-    std::sregex_token_iterator end;
-
-    std::vector<std::string> out;
-    for (; iter != end; ++iter)
-      out.push_back((std::string) * iter);
-    return out;
-  }
-
-  // gets the shape of a 2D vector
-  template <typename T>
-  std::vector<int> get_poses_shape(std::vector<std::vector<T>> arr) {
-    std::vector<int> out;
-    for (auto i : arr)
-      out.push_back(i.size());
-    return out;
-  }
-
   // convert vector of strings to type
   template <typename T>
-  std::vector<T> split_to_number(std::vector<std::string> split)
+  inline std::vector<T> split_to_number(std::vector<std::string> split)
   {
     // converts a vector of strings to a vector of doubles
     std::vector<T> out(split.size());
@@ -138,7 +107,7 @@ namespace SockReceiver {
   }
 
   // something socket related idk
-  void remove_message_from_buffer( char* buf, int& numbytes, int msglen )
+  inline void remove_message_from_buffer( char* buf, int& numbytes, int msglen )
   {
     // remove complete message from the buffer.
     // thanks to https://stackoverflow.com/a/13528453/10190971
@@ -146,33 +115,6 @@ namespace SockReceiver {
     numbytes -= msglen;
   }
 
-  // char buffer to string
-  std::string buffer_to_string(char* buffer, int bufflen)
-  {
-    std::string ret(buffer, bufflen);
-
-    return ret;
-  }
-
-  // splits string somehow
-  std::vector<std::string> split_string(std::string text)
-  {
-    std::istringstream iss(text);
-
-    std::vector<std::string> tokens{std::istream_iterator<std::string>{iss},
-                            std::istream_iterator<std::string>{}};
-
-    return tokens;
-  }
-
-  // yes
-  bool strings_share_characters(std::string a, std::string b)
-  {
-    for (auto i : b) {
-      if (a.find(i) != std::string::npos) return true;
-    }
-    return false;
-  }
 
   class Callback {
   public:
